Refactor float to int for frame and timing variables
Changed several variables from float to int for frame counts and timing, including chargeTiming in Enemy4.cpp and kMayuFrames in Shrinemaiden.h. Also updated random value assignments to use static_cast<int> for consistency. Minor fix in ModeSelectScene.cpp to use float literal for brightness. Updated Stage10.cpp to use static_cast<float> for random spawn positions.

+++ DX11_hew2/Enemy4.cpp
@@ -116,7 +116,7 @@ void Enemy4::move()
 				chargeTimer = 0;
 
 				// ŽŸ‰ñ‚Ì‘Ò‹@ŽžŠÔ
-				chargeTiming = (float)get_rand_range(2, 4);
+				chargeTiming = static_cast<int>(get_rand_range(2, 4));
 
 				chargeTarget = miko_pos;
 				chargeTarget.z = 0.0f;
@@ -163,7 +163,7 @@ void Enemy4::move()
 					// ‘Ò‹@‚Ö–ß‚·
 					isCharging = false;
 					chargeTimer = 0;
-					chargeTiming = (float)get_rand_range(2, 4);
+					chargeTiming = static_cast<int>(get_rand_range(2, 4));
 
 					m_velocity = 0.0f;
 					stunTimer = kHitCooldown; // ˜A‘±ƒqƒbƒg–hŽ~
@@ -219,7 +219,7 @@ void Enemy4::move()
 				{
 					isCharging = false;
 					chargeTimer = 0;
-					chargeTiming = (float)get_rand_range(2, 4);
+					chargeTiming = static_cast<int>(get_rand_range(2, 4));
 
 					m_velocity = 0.0f;
 					stunTimer = kHitCooldown;
@@ -242,7 +242,7 @@ void Enemy4::move()
 			{
 				isCharging = false;
 				chargeTimer = 0;
-				chargeTiming = (float)get_rand_range(2, 4);
+				chargeTiming = static_cast<int>(get_rand_range(2, 4));
 
 				m_velocity = 0.0f;
 			}
